+++ docs/CHANGELOG.md
@@ -10,6 +10,17 @@ In summary, [SemVer](https://semver.org/) can be viewed as ` Breaking . Feature 
 
 **Version `0.x.x` doesn't have to apply any of the SemVer rules**
 
+## Version 0.1.1 2020-07-07
+
+- Fix XSLT locations to 1.3 specs, see <https://www.gob.mx/sat/acciones-y-programas/contabilidad-electronica-173700>.
+- Test source string after sign documents.
+- Add exception when empty source string.
+- Use CSD instead of FIEL to sign documents.
+
+Development:
+
+- Change `php-cs-fixer` to version 3.
+
 ## Version 0.1.0 2020-01-01
 
 - Initial working release for testing with friends